Bug fix with the inital xmlRepository folder being empty.
On corporate networks, sometimes after Windows Updates Osprey renders off the screen and minimized to a tiny box. You can't see it. Manual remedy is to hover over the its taskbar icon, right click on the resulting thumbnail, and select "Maximize".
But, this commit checks the window size and location in the Form Shown event, moves the form to (200,200), and width and height to 500x500.